updated calculation of relative branch distance
Previously relative distance along subbranch was calculated for cylinder midpoint but now the endpoint is used.

This change allows for was made to allow for relative position values of 1. Also, since next a cylinder length preprocessing is implemented to the direct method (previously only LCL method), the endpoint of the cylinder should be good enough reference point for the cylinder.

There is still an easy option to change the code to use the midpoint if needed.