leaf size sampling modified for QSM direct
Now a leaf is sampled on cylinder if there is leaf area budget for at least one leaf (as calculated by mean area of leaves). Previously it was enough if the cylinder had any leaf area budget, which resulted in almost all of the cylinders getting at least one leaf.

This threshold might, however, be too high and cause the LADD distancewise distribution to be skewed to the right towards larger values. Thus there might be a golden middle road to be chosen here.