added cylinder preprocessing for QSM direct
Now the QSM cylinders are preprocessed before calculating leaf area budgets such that if a cylinder length ratio is more than 0.1 compared to total length of the subbranch, it is cut into smaller cylinders. This allows for more accurate leaf area budgets with respect to relative branch distance (and naturally also relative height and compass direction).

Previously, small branches with only a few or even only one cylinder presented too coarse discretization for sampling of distancewise LADD, which became apparent in a QSM that had most of its higher order branches containing only one or two cylinders. The distancewise LADD is sampled from the end point of the cylinder, so in these cases the whole of the branch would get way too much leaf area budget.

Also, the visualization of distancewise LADD has been modified to be more accurate: now the twig start point is projected to cylinder axis and this position is used to accumulate leaf area in each distance bin.