removed unnecessary polar integration
In the function that calculates leaf area budgets, the integration of compass direction target function was done with polar integration, which is not necessary in this case (as far as I suppose...). That is because we want to allocate specific amount of leaf area to specific directions and follow the defined target distribution, which is actually defined for a linear compass coordinate instead of circular compass coordinate.