tuning visualization and compass dir calculation
Modified some plotting parameter for all of the marginal distribution of LADD.

The reference origin for calculating compass direction is now the mean position of cylinders weighted with cylinder length. Previously all the cylinders had the same weight, in which case a lot of short cylinders on one side of the tree could pull the origin to a non-intuitive position.