removed bug from library generation
Initially it was assumed that the discretization of cylinder inclination angle starts from 0 and ends in pi. This is, however, not smart since it makes more sense to start the discretization somewhere other than the endpoints.

This original assumption caused that the first and last inclination indexes for cylinders would cause vertical cylinders, for which the azimuth angle has no meaning in uniform petiole direction sampling.  Since this is not necessarily the case anymore, and for phyllotaxis and petiole direction distribution this has effect, the condition has been modified.