leaf cylinder library overhaul
Parallelized the leaf cylinder library generation and changed the data type to be a struct instead of a cell.

The benefits of parallelization are obvious. The change from cell to struct removes the need to use dummy dimensions in the library, as cell would omit dimensions with only one node value and thus requiring another dummy value to keep indexing ok. Referring to library cylinders in now done with ind2sub and sub2ind functions.