bug fix for phyllotaxis
The petiole orientation definition when using phyllotaxis calculated the petiole direction wrong.

Also, previously the petiole inclination angle with respect to cylinder axis was dependent on the LOD, but now the default value is pi/2 and can be set to some other by the user. Setting this angle to a constant value instead of letting it be free ensures a stricter phyllotaxis pattern.