generalize point cloud handling
Now in canopy hull method the point cloud can have any z-values instead of needing to have the lowest point exactly at z=0 plane. This makes the function generate_foliage_canopy_hull a lot more general as the point cloud can now have any coordinate values if the stemCoordinates is supplied or the point cloud can be at any height value if the stem center is assumed to be the z-axis.