small updates to ch code
Now the 'HoleTreshold' parameter of alpha shape generation is set to 10, meaning open volumes smaller than 10 m^3 are included in the alpha shape (the value of 10 chosen arbitrarily here). Without this, there could be small empty volumes within the shape which can intersect with the assumed stem center and therefore prevent leaves from being generated on certain heights.

In addition, the commit contains few bug fixes to the code.