expand point cloud sampling functionality
Now the point cloud sampling can be done also in the case where all three LADD marginal distributions are defined. In this case, when point cloud sampling is used, the heightwise LADD is sampled for the leaf height value and then this is used to either use point cloud sampling on that height or alternatively the other two marginal LADDs are used to define leaf start point (depending on the PCSampling weights).

The commit also has minor bug fix.