add qsm-based leaf positioning
A new feature for leaf positioning: instead of parametric LADD, the user can choose to position the leaves "uniformly" around the QSM cylinders. It is strictly not uniform as:
- stem cylinders get no leaves
- the leaves positions are weighted towards the ends of subbranches
- the leaf area per cylinder is scaled with cylinder length
The idea is to be able to generate foliage where the leaves are automatically spread around the tree in somewhat of a realistic manner. This feature is implemented to allow for generating foliage on trees when exact shape of LADD is not of big importance. For example, this is useful when adding leaves on a collection of QSMs of different shape and size to create a virtual forest plot.